requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bleach==6.3.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
cloudpickle==3.1.2
contourpy==1.3.3
cycler==0.12.1
fastapi==0.136.0
filelock==3.25.2
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kaggle==2.0.1
kagglesdk==0.1.19
kiwisolver==1.5.0
lightgbm==4.6.0
llvmlite==0.47.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
multidict==6.7.1
narwhals==2.19.0
networkx==3.6.1
numba==0.65.0
numpy==2.4.4
packaging==26.1
pandas==3.0.2
pillow==12.2.0
plotly==6.7.0
propcache==0.4.1
protobuf==7.34.1
psutil==7.2.2
pyarrow==23.0.1
pydantic==2.13.2
pydantic_core==2.46.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-louvain==0.16
python-multipart==0.0.26
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
setuptools==70.2.0
shap==0.51.0
six==1.17.0
slicer==0.0.8
starlette==1.0.0
sympy==1.14.0
text-unidecode==1.3
threadpoolctl==3.6.0
torch==2.11.0+cpu
torch-geometric==2.7.0
torch_scatter==2.1.2+pt211cpu
torch_sparse==0.6.18+pt211cpu
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
webencodings==0.5.1
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
